refactor rates comments
