+++ src/admin/redux/reducer.js
@@ -5,7 +5,7 @@ import payment_defaults from './modules/payment_defaults.js'
 import mode from './modules/mode.js'
 import payments from './modules/payments.js'
 import charge_form from './modules/charge_form.js'
-import route from './modules/route.js'
+import route from '../../shared/redux/modules/route.js'
 import payment_reports from './modules/payment_reports.js'
 import letter from './modules/letter.js'
 import email from './modules/email/reducer.js'
