+++ src/admin/routes.js
@@ -17,7 +17,7 @@ import Emails from './containers/email.js'
 import Labels from './containers/labels.js'
 import Letters from './containers/letters.js'
 
-var { pathDidUpdate } = require('./redux/modules/route.js')
+var { pathDidUpdate } = require('../shared/redux/modules/route.js')
 
 hashHistory.listen((path) => store.dispatch(pathDidUpdate(path)))
