+++ src/shared/redux/middleware/heartbeat.js
