+++ src/user/user_app.js
@@ -1,17 +1,14 @@
 import React from 'react'
 
 import NavBar from './containers/nav_bar.js'
-import App from '../shared/app.js'
 
 export default class UserApp extends React.Component {
   render () {
-    const add_details = this.add_details || id
+    const clone_child = child => React.cloneElement(child, { name: this.props.name })
     return (
       <div>
         <NavBar />
-        {React.Children.map(this.props.children, add_details)}
+        {React.Children.map(this.props.children, clone_child)}
       </div>
   )}
 }
-
-function id (arg) { return arg }
